requirements.txt
curl_cffi
beautifulsoup4
python-dotenv
requests
quart
patchright
rich
